requirements.txt
flask
redis
boto3
python-dotenv
gunicorn
requests
pytest
Pillow
pillow-heif